postgres: only connect to read-write nodes
If several nodes are specified (using comma separated values in the
host), zabbix will try to connect to the first one and if it is a
replica, it will fail.

With this new parameter, it will continue trying the other servers on
the list until it connects to a read-write node.